test(fms): k2 收款核销链 + k3 过度核销防护
验证 cash_journal create/confirm（依赖 055 seed）+ write_off 核销与
未核销余额；过度核销累计超 source_total 报 OverWriteOff。

+++ abt-core/src/fms/write_off/repo.rs
@@ -190,8 +190,8 @@ impl WriteOffRepo {
         source_id: i64,
     ) -> Result<Decimal> {
         sqlx::query("SELECT pg_advisory_lock($1, $2)")
-            .bind(source_type.as_i16() as i64)
-            .bind(source_id)
+            .bind(source_type.as_i16() as i32)
+            .bind(source_id as i32)
             .execute(&mut *executor)
             .await?;
 
@@ -212,8 +212,8 @@ impl WriteOffRepo {
         source_id: i64,
     ) -> Result<()> {
         sqlx::query("SELECT pg_advisory_unlock($1, $2)")
-            .bind(source_type.as_i16() as i64)
-            .bind(source_id)
+            .bind(source_type.as_i16() as i32)
+            .bind(source_id as i32)
             .execute(&mut *executor)
             .await?;
         Ok(())
